requirements.txt
# PyTorch 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
Pillow>=9.5.0

# 模型信息显示
torchinfo>=1.8.0

# 进度条显示
tqdm>=4.65.0

# 可视化
matplotlib>=3.7.0
